+++ uebung1/IEvent.java
@@ -1,4 +1,4 @@
 public interface IEvent {
 	void eventExec(BasicSimulator simulator);
-	double getExecTime();
+	int getExecTime();
 }
